backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
torch==2.4.1
torchvision==0.19.1